% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myq_euclidean.R
\name{myq_euclidean}
\alias{myq_euclidean}
\title{Euclidean Distance}
\usage{
myq_euclidean(zm, point = c(1, 2, 1))
}
\arguments{
\item{zm}{Input matrix of set of vectors.}

\item{point}{Coordinates of the point to calculate distance to the input.}
}
\value{
Sum of the euclidean distance from point to set of vectors.
}
\description{
Function to calculate euclidean distance between two vectors.
}
\examples{
simulate.outer(n=1e3, d=3, q.outer=myq_euclidean, point=c(1,1,1))

}
